Don't save curModule in module parser
The `module` parser was previously used for path parsing. It is
currently only used for parsing module prefixes of identityref leaf data
values. However, that is actually a mistake. Before the leaf_data parser
rework, there was a another module parser, that DIDN'T SAVE the
curModule (which is the point of this patch), but I accidently deleted
the wrong one and used this one, that saved the current module. This led
to a bug, where the identityref sets the current module, and then the
list called "inventory", in the new test, gets parsed as
"other:inventory".

Also, I'm gonna get rid of curModule, so one less usage of the variable
makes it easier.

Change-Id: I828750092c7fba71c1646e3aa6c190d9e07c58af