Simplify presenceContainerPath_class
So... this might not seem as such a good change, since the result is an
increase in lines of code. But! This really is a cleanup from old-style
path parsing. There is a reason most of the methods in Schema have two
arguments: "location" and "node". This is because before, when I wanted
to check what node type a path is (when parsing), I only had the string
with the node's name. However, this was never really a case in
presenceContainerPath_class and the old code had to do weird stuff to
convert the format Schema accepts.

I'm doing this change separately, because it kind of is a big deal (the
parser gets changed), but ideally I'd like to change all of the methods
to have only one argument.

Change-Id: Ib38901f3c018f4b5a72bf31ce376b68d030cc13b