Fix .. when inside a path with no child nodes
The ".." string was being added to the symbol table on each iteration of
availableNodes. That means, if there were no child nodes in the current
context path (the one stored in Parser), no ".." was added, so you
couldn't go up. It was also somewhat inefficient to add it on each
iteration.

Change-Id: If9e17a545bd175843d8d0a8897ffeb74c07de0c7