Fix get interpreter
The previous algorithm advanced the iterator after printing a leaflist
instance, which meant that if the last item was leaflist instance, the
iterator would be advanced to .end() and in the next while-loop
condition check, the iterator would no longer be dereferencable.

The new algorithm peeks at the next element, checks whether it's the
end, if not, checks whether it's still the same leaflist and only then
advances the iterator.

Issue: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/175
Change-Id: I0770d87585da800c160684221a85dbeb687a8955