Rework absolutePathFromCommand
The problem with absolutePathFromCommand was that in its template
version, it could only work with some types of paths, and if the
template didn't match, another overload was used. This means that any
change to path processing would need to be done more times.

A new function called resolvePath (which admittedly isn't the best
name, but I think it's better) now works with all types of paths.

Change-Id: I85f2eb4ba2bccaeb996cd60d47dac0ae1a6b43a3