Add test for getting "/"
When playing with yang-cli, I found that just putting in "get" didn't
return any data, even though some data was available. All of the
DatastoreAccess implementations handle "/" (getting everything) in a
special way, so I added a test which tests if "/" works at all (the bug
was that yang-cli didn't handle this special case correctly).

Change-Id: I9052e0f33da43f6b8158964b4eb269f8d4a2064a