tests: Accessing non-existing schema nodes
This is really hilarious because apparently sysrepo's behavior isn't
terribly consistent. So I decided to make the maybe-except-exception
wrapper a tad more generic.

Of course there was an ugly error because now that I'm templating on the
expected exception, I cannot use REQUIRE_THROWS_AS because that one is
written with an implicit assumption that the exception name is not a
type-dependent name. This means that (AFAIK) I have to enumerate all of
these possible exception types explicitly.

The final catch was that type-dependent static assert. After that one
everything was 🌈 and I felt 🥰 and saw nothing but 🦄.

Change-Id: I8862b50bf46cd156e71c83640d9d668989b4db9d