Remove default destructors from exceptions
I think there is no reason to explicitly define default
destructors. Honza's comment on this patch:
"So this has been something that I've been doing for years, and the
reason for doing that is that I wanted to make sure I can throw/catch
exceptions over DSO boundaries. It might have been while the dinosaurs
were still ruling the Earth and GCC 3.4 was the new king of the block,
and people were playing with -fvisibility=hidden. This means that
nothing of that might apply today, but it also means that I do not
understand whether it is still an issue or not.

The way I remember stuff was that the "first virtual"'s definition drove
vtable generating, and that determined the type_info. A TL;DR version is
that a class like this one, with no explicit definition of a virtual (or
a first virtual?), caused at least two TUs in two different DSOs to
contain "a definition" of the exception's
type_info/RTTI/whatever-it-is-called. As a result, if a library code
that eventually ended up in libfoo.so did this:

 void bar() {
  // ... something useful here
  if (die)
    throw my_exception;
 }
and the main app (or perhaps a different DSO?) did this:

 void another() {
   // ...
   try {
     bar();
   } catch (my_exception&) {
     // THIS WAS NOT BEING CALLED because the two sides had two separate copies of the type_info for my_exception
   }
Now,
https://stackoverflow.com/questions/10787492/exceptions-across-binary-boundary
looks relevant, and indeed, apparently it's been fixed/changed since GCC
4.5: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=47960 . Reading other
bits such as
https://dholmes215.github.io/c++/2017/06/30/fun-with-typeid.html brings
these memories back.

So, if you're willing to vouch that removing these "just so that Honza
feels good" destructors that are =default in the .cpp file, that this is
a good and safe thing, please do that in a separate commit, and please
copy-paste this comment from me there. The risk that I'm worried about
is "what's gonna happen when we refactor this to use a DSO for sysrepo
and another DSO for netconf".

If the bug Honza described ever appears, I can readd the destructors.

Change-Id: I287b38f469db699c8510ed59c788cc2b830a8560