Fix enum key parsing
I'm not sure if this got broken during leaf_data rework, or if it was
always broken, but enum parsing had to get reworked.

The problem was parsing enums with "+char_". This works with set,
because set doesn't have any additional arguments. However, that's not
the case with key parsing. In key parsing +char_ was eating too much of
the input (mainly the closing bracker). At first I tried something like
"+char_-']'", but enums are normal strings, so ']' is a valid character.
That's why I reworked the parser to just try all the possible enums.

Change-Id: I5996286424dd95377b55f90eeba29d9a5eba731c