Update patch set 1
Patch Set 1:

It seems that this uses yet another approach compared to the other path parsers in the grammar (doesn't it?). My preference would be to follow the same pattern for all paths (and that probably means doing it in a manner similar to cdPath_def, if possible). Will that work?

Patch-set: 1
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>