Fix leafDataFromValue
Leafrefs were not working correctly, when pointing to a data node not in
the tree. Because of this I had to rework the thing a little bit (I
basically copied the algorithm from cla-sysrepo, where I had the same
issue). I also changed the name. And also added test, both versions of
leafref included.

Change-Id: I8a102c54bf58a4221610c8a83d26979fb7b4461d