fix linking on ARM
Here's what I'm getting from Buildroot on a real ARM HW:

 ~ # sysrepo-cli
 sysrepo-cli: error while loading shared libraries: liblibyangutils.so: cannot open shared object file: No such file or directory
 line-NFP ~ # ldd $(which sysrepo-cli) | grep yang
         liblibyangutils.so => not found
         libyang-cpp.so.1 => /lib/libyang-cpp.so.1 (0xb6dd6000)
         libyang.so.1 => /lib/libyang.so.1 (0xb6ccd000)

Fix this by making sure that all our libraries are produced STATIC, not
dynamic. Also, do not try to use library names prefixed with "lib"
because that's a prefix which is very often omitted. The fewer bits that
could go wrong the better, IMHO.

Fixes: 2e4cafe3 ("Fix leafDataFromValue")
Change-Id: I768e93d76cad6817f98fdecc925ce2151bece7c1