Add test for lyNodesToTree
While working on key-less lists, I realized there was no test for this
function. So, now there is one and it includes key-less lists. This test
doesn't increase coverage, because there's no special code for handling
key-less lists. Libyang deals with them in its own way, by assigning
indices to their XPath, so this tests checks for consistency of that.

Change-Id: I2d70436a024198c0a16534e42425c57adc7c15ab