Make sysrepo dependency optional
This project is useful as a standalone NETCONF client, too. Do not
require sysrepo unconditionally. If not available, just build the main
NETCONF client (and a `yang-cli`, too, because it's essentially free).

The CMake setup is a wee bit complex. CMake's OPTION statement doesn't
really support tri-state dependencies, and I like automagic stuff:

- OFF for disabling stuff unconditionally, even if the deps are there,
- ON for enabling stuff, and failing if it cannot be built,
- AUTO for trying to enable stuff if possible, and gracefully disabling
  it if not available

Of course this has a potential of introducing nasty surprises to the CI,
so make sure the CI fails if the sysrepo backend gets disabled for some
reason.

Change-Id: I9c38245c3eea767ad20f42240328c483b6f80dd4