Update patch set 4
Patch Set 4:

(1 comment)

> Sorry, I couldn't help it, but try my own take on this: https://gist.github.com/syyyr/d46657d39262fabaffe112d5bc8b393e
> 
> I made the stuff a little flatter, with less nested ifs, but I'm not entirely sure, if I preserved the same functionality. Let me know what you think.

I've looked at the diff, and it seems that the most important part is moving the code which only fires when sysrepo is found (and should have been searched for) to a top-level if statement. I've implemented that.

I've also done some variable renaming, nothing important. Hopefully DO_* should hint that it's an internal thingy.

Patch-set: 4