Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> Oh, nice:
>  
> ==2078==ERROR: AddressSanitizer: odr-violation (0x000000e7bc00):
>   [1] size=96 'nacm' /home/ci/src/cesnet-gerrit-public/CzechLight/dependencies/libyang/src/extensions/nacm.c:157:26
>   [2] size=88 'nacm' /home/ci/src/cesnet-gerrit-public/CzechLight/dependencies/Netopeer2/src/netconf_acm.c:31:13

OK, the `nacm` in that libyang plugin looks like something that's tied to plugin workflow (I kinda assume that in a library nacm.so, there's a global variable with "visible linkage" called `nacm`). You probably cannot rename that.

In Netopeer2, OTOH, it looks safe to just change that variable in src/netconf_acm.c to make sure it's `static` (hopefully that should do the right thing re ODR). Or just rename it (but making it a static won't hurt anyway).

Patch-set: 7