Update patch set 2
Patch Set 2:

(2 comments)

The first thing which sticks out as a sore thumb are the m_ prefixes. I know I used them in cla-sysrepo for private members, but here they are used as a part of the "public API". Perhaps the code would be cleaner when the m_ prefix is removed?

I like the change of boost::process because that's an external class. OTOH, I'm not sure that the changes to completions produce that much value.

Patch-set: 2
Reviewer: Gerrit User 1000000 <1000000@e32f8df8-3db0-4f76-a6c6-fe9d6e69dd68>