tests: Shorten paths to the NETCONF listening socket
On any Unix, there's a limit on the maximal length of a UNIX socket
path. Linux defines that to 108 characters (`UNIX_PATH_MAX` in
`<linux/un.h>`, [1]), and of course that was too long for my particular
build setup:

 >>> len('/home/jkt/work/prog/_build/czechlight-clang-11/netconf-cli/test_netopeer_files/test_datastore_access_netconf.sock')
 113

Because the build dir prefix is outside of our control, let's just use
a relative path. Tests are supposed [2] to run from the
`CMAKE_CURRENT_BINARY_DIR`, so this should be safe.

The awesome bit about this is that it's actually "mostly OK" to use
these longer path names because they appear to be silently truncated.
The tests still run, a socket gets created, and the test manages to
connect just fine. The failure is only visible when a `*_cleanup` test
runs which tries to `rm` a non-existing path.

[1] https://github.com/torvalds/linux/blob/master/include/uapi/linux/un.h
[2] https://cmake.org/cmake/help/latest/prop_test/WORKING_DIRECTORY.html#prop_test:WORKING_DIRECTORY

Change-Id: Ibd96fc779d0ec6174e081b41004a15a6afb8eb35