NETCONF: do not try to use confirmed-commit
It is not supported by our server, and in any case it would be an error
to blindly rely on an optional feature of the protocol. On the other
hand, if we ever support some capability autodiscovery *and* the user
sets a pref or an option for a confirmed commit, that would be a good
place to turn this feature on.

Also, libnetconf2 started enforcing this last week, so make sure we
continue working with it.

Change-Id: I01dc4a324bdd91bc59929a540cdbfaff0ab0dc1f
Fixes: https://github.com/CESNET/libnetconf2/commit/283c3c0f16f63a20b7f183a35cbfc4d30457f1ea