NETCONF: do not try to use confirmed-commit
It is not supported by our server, and in any case it would be an error
to blindly rely on an optional feature of the protocol. On the other
hand, if we ever support some capability autodiscovery *and* the user
sets a pref or an option for a confirmed commit, that would be a good
place to turn this feature on.

Change-Id: I01dc4a324bdd91bc59929a540cdbfaff0ab0dc1f