Fix sending actions through ProxyDatastore
DatastoreAccess::execute functions always expect input with the prefix
deleted. This does make the test somewhat prettier, because you don't
need to repeat the prefix all the time, however, this adds complexity to
the `execute` functions, because you have to consider that. Also what
really happens inside NetconfAccess/SysrepoAccess is that it just
concatenated again.

The actual bug is this: when using ProxyDatastore datastore to send
RPC/action, getItems is called to get the input. This algorithm was
supposed to strip the prefixes. But there was an error in this stripping
algorithm, because stripping an "action" path is more difficult than RPC
because it can be nested.

The solution: get rid of the compacted paths in input to simplify the
algorithms.

Change-Id: I28acaffadb55d89e508d75d58d365a3523f295bb