Get rid of trailing slash saving
This m_trailingSlash was the source of bugs, so I got rid of it. A
caveat is that I lose this information if I want to recreate the path
back, but it isn't really too useful.

The actual bug was about conversion of the parsed path back to a string.
The problem was when the input had a trailing slash, the result string
would also have that trailing slash, which is undesirable (libyang
cannot deal with it).

This patch completely removes the information about the trailing slash,
so converting a *Path_ structure back to a string can't result with a
path that has a trailing slash.

However, I do think that there should be at least some sort of a test,
so the `cd` test converts paths back to strings and asserts that they
don't have trailing slashes.

Issue: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/212
Change-Id: I08e02401580ce31c1e0412a5798cea20e7802ab4