Fix `switch` parsing
For some reason, the completion generator prevented the symbol table
parser from running and the value inside the resulting switch_ was
undefined. I THINK this has something to do with the fact that switch_
is a one member structure and that's something that Spirit doesn't
support (or at least it didn't last time I checked). Asserting the
attribute of the completion generator seems to solve the issue.

I do think my use-case for using a one member structure is valid (it's
because I want to differentiate commands by types and they have the same
number of arguments), so I think I'll ask upstream about this.

Issue: https://tree.taiga.io/project/jktjkt-netconf-cli/issue/218
Change-Id: I364b32b76741c198808cc2b3c5027913162d0703