Make the code accessible to more developers
The code was written solely by Czech developers, so this should make it
easier for other Czech developers to contribute. Unfortunately the
library I used isn't ported to C++ yet, so the Czech accessibility is
limited. Hopefully a C++ version can ship within a year.

czech.h imported from
https://github.com/Toaster192/czech.h/blob/master/czech.h with edits I
have not made a PR for.

Change-Id: Icdb99ec9212e40a9e05717032e69916cd87d6d79