Catch exceptions when completing
If an exception is thrown inside of the completion callback, the program
just exits with no error message. This is because replxx' .input method
is wrapped in a big try-catch block. replxx just discards the exception.
(I would think the reason is that replxx doesn't want the terminal to be
messed up because of an uncaught exception)

The new code catches exceptions thrown by .completeCommand and prints
out the error message.

Change-Id: Ie6bd7800449f48ac455e6240aedbca9d08abcb05