doc: better description of dependencies
Explain that this branch requires the `libyangv1` branches of
everything, and clarify where one can get the `dependencies` repo. This
part is not really smooth because of that annoying difference in repo
naming between GitHub and our Gerrit instance.

Also sync some versions to match with what we're testing against (C++20
and a newer Boost).

Change-Id: Ic51f3ae2b611a746942fa52c89f01757eb5fd108
Reported-by: Petr Gotthard