Rework x3::phrase_parse into x3::parse
For commands, there is predefined places where we want to skip spaces:
on the start of the command, in between command arguments and at the
end of the command. Using phrase_parse just makes everything more
complicated, since I have to actually disable space skipping a lot of
the time.

Change-Id: I90b32e1a270da0c228691935cc7206d73248025a