CI: stop pinning submodules/dependencies in git
Traditionally, we've been pinning the NETCONF stack via a separate git
repo which was added as a git submodule for `cla-sysrepo` and
`netconf-cli`. This has worked well, but it caused a certain level of
overhead, and it led to possible inconsistencies (e.g.,
`CzechLight/br2-external` has code which tries to guesstimate whether to
pick the submodules from `cla-sysrepo` or `netconf-cli`).

Also, the other repos (`velia` and `rousette`) were always using the
latest master of `CzechLight/dependencies`, and it has worked well.

Change-Id: Ie9bf475c446c4330105c7631f2ab63a0358ca605