Fix PathParser attribute_type
This took me a while... For some reason, Spirit does not really NEED
this `attribute_type` thing to work properly. Or maybe it does, but
doesn't care if it's invalid, because obviously, `::type` has to be
used, otherwise the using is just the ModeToAttribute type.

Anyway, fixing this revealed another bug on the code: until now, the
getPath parser used decltype(get_::m_path) as the attribute, which is
`boost::optional<boost::variant<dataPath_, module_>>. However, this is
not the correct attribute, because the getPath parser can only parse a
dataPath_, or a module_. The "optional wrapping" is done via the
optional parser directly in the rule for the get command.

The reason I found all of this is that I wanted add an optional argument
for `get` and I kept getting errors for "can't assign
boost::optional<boost::variant<...>> to boost::variant<...>".

Change-Id: I76ff8d3f201b58bb83b68c428f6b95f5c87f3a75