Add a function for staticSuggestions
This new function wraps the boilerplate that's needed for making static
suggestion list. This could also be used for the dump_args parser (to
get rid of more boilerplate), but it's a bigger change, so I'll leave
that for another patch.

Also, the parser really NEEDS its attribute to be the unused_type,
otherwise weird stuff happens, ghosts appear and it also messes with the
attributes of other parsers. I don't know why, but oh well.

Change-Id: Ida0bcc86124b88d686ac86cf4743d220db86f59b