Move rule tag definitions
Spirit's x3::rule needs some class type so that it can recognize rules.
This class can have some on_success handler which adds some more logic
to the parser. However, some of the classes don't have on_success
handlers. That means that they can be directly declared in the
x3::rule's definition. This indicates that there's the reader of the
code doesn't have search for the definition (and check whether there' an
on_success handler).

I'm pretty sure that this can't really be enforced, so really, it's just
a convention.

Change-Id: Ib6e5e76174ad19737242dd8c074a0bbc867e195c