Rework lyNodesToTree
The original algorithm is very old and was designed for old libyang. I
have completely scrapped it and made a new one.

lyNodesToTree iterates over the input collection and uses DFS over each
element. The new algorithm now merges this iteration. The advanceIt()
function handles both the inner and the outer iteration.

Change-Id: I0c8d4059439254f69ac056ee9941dcfe1cc63cd2