CI: allow longer builds
This projects regularly takes 10-15 minutes to build, using the current
generation of build VMs. That's also our default timeout which works
well for the other projects, but it sometimes breaks this one.

Since we're usually setting the time to about twice the regular
duration, let's bump it to half an hour here.

Change-Id: I0a379806e33b893800569e1b340630ab95313cfe