Fix build with Boost 1.78 and 1.79
Up to Boost 1.77, it was enough to just call this from the `on_success`
handler:

 _pass(context) = false;

...and the spirit-x3 core would auto-revert the iterators. The sky was
pink and the unicorns were flying with their horns upright. Starting
with 1.78, though, the auto-reposition of the input iterators was
removed, which necessitated resetting them by hand. However, this
required casting away the constness of the `after` iterator because the
function signature was just wrong, passing a non-const iterator
reference for the `before` iterator (which we do not need to modify),
and a const one for the `after` iterator that we do want to modify.
Functions with a non-const reference to the `after` iterator wouldbe
silently ignored ("hey, it's just a function that I'm not calling, no
***** given").

That was fixed in 1.80, but we want to make sure that we work on the
latest Fedora which only has Boost 1.78, even though upstream already
released Boost 1.82.

Fix that by using `const` function signatures, but cast that const away
at runtime.

Also for consistency, always overwrite that iterator whenever rejecting
the input -- no matter whether the current version of Boost needs it or
not, and even in cases where the previous version of code would "just"
reject parsing, but keep the after iterator on its new value.

Bug: https://github.com/boostorg/spirit/pull/711
Bug: https://github.com/boostorg/spirit/issues/703
Fixes: 5b8e2825 Fix for Boost 1.78
Change-Id: Ib603db643a086e0351a4d74082505329ab8944cb